Refactor from PR feedback
